// src/api/model/updateContactDto.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * accounting
 * Accounting API description
 * OpenAPI spec version: 1.0
 */

export interface UpdateContactDto {
  id: string;
  name?: string;
  country?: string;
  email?: string;
  phone?: string;
  street?: string;
  city?: string;
  psc?: string;
  ico?: string;
  dic?: string;
  description?: string;
}